refactor: выделяет сортировку в отдельный компонент
* Добавляет кастомный хук useOutsideClick
