+++ src/shared/lib/react/index.ts
@@ -3,5 +3,6 @@ export { PrivateRoute } from './components/privateRoute';
 
 export { useAppDispatch } from './hooks/useAppDispatch';
 export { useAppSelector } from './hooks/useAppSelector';
+export { useOutsideClick } from './hooks/useOutsideClick';
 
 export { AuthorizationStatus } from './const';
